% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extracDATA.R
\name{extracDATA}
\alias{extracDATA}
\title{Title}
\usage{
extracDATA(user = "SEI", variable = "PREC", station.cod = "21010320")
}
\arguments{
\item{user}{}

\item{variable}{}

\item{station.cod}{}
}
\description{
Title
}
