% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screeningDATA.R
\name{screeningDATA}
\alias{screeningDATA}
\title{Title}
\usage{
screeningDATA(vardata, typevar, screen.method)
}
\arguments{
\item{vardata}{}

\item{typevar}{}

\item{screen.method}{}
}
\description{
Title
}
